web/requirements.txt
django 
pillow
django-admin-interface
django-crispy-forms
django-rest-framework
django-pwa